Configure caching for the getAllProducts function
