+++ firebase.json
@@ -16,7 +16,7 @@
       "**/.*",
       "**/node_modules/**"
     ],
-    "rewrites": [   
+    "rewrites": [
       {
         "source": "/api/allProducts",
         "function": "getAllProducts"
