+++ functions/package.json
@@ -15,8 +15,10 @@
   },
   "main": "lib/index.js",
   "dependencies": {
+    "@types/cors": "^2.8.10",
     "@types/nodemailer": "^6.4.0",
     "axios": "^0.21.1",
+    "cors": "^2.8.5",
     "express": "^4.17.1",
     "firebase-admin": "^9.2.0",
     "firebase-functions": "^3.11.0",
